+++ packages/desktop/components/modals/ProfileActionsMenuModal.svelte
@@ -8,17 +8,7 @@
     import { checkOrUnlockStronghold } from '@core/stronghold'
     import { diffDates, getBackupWarningColor, isRecentDate } from '@core/utils'
     import { PopupId, closePopup, openPopup, popupState } from '@desktop/auxiliary/popup'
-    import {
-        Button,
-        ButtonSize,
-        DeveloperIndicatorPill,
-        Icon,
-        Modal,
-        ProfilePicture,
-        Text,
-        TextType,
-        Toggle,
-    } from '@ui'
+    import { Button, ButtonSize, DeveloperIndicatorPill, Icon, Modal, Avatar, Text, TextType, Toggle } from '@ui'
     import { fade } from 'svelte/transition'
 
     export let modal: Modal = undefined
@@ -86,7 +76,7 @@
 >
     <profile-modal-content class="flex flex-col" in:fade={{ duration: 100 }}>
         <div class="flex flex-row flex-nowrap items-center space-x-3 p-3">
-            <ProfilePicture profile={$activeProfile} size="small" />
+            <Avatar profile={$activeProfile} size="small" />
             <div class="flex flex-row items-center space-x-2">
                 <Text>{profileName}</Text>
                 {#if $activeProfile?.isDeveloperProfile}
