+++ packages/shared/src/components/Avatar.svelte
