+++ packages/shared/src/components/index.js
@@ -1,5 +1,6 @@
 export { default as AppUpdateToast } from './AppUpdateToast.svelte'
 export { default as Animation } from './Animation.svelte'
+export { default as Avatar } from './Avatar.svelte'
 export { default as Button } from './Button.svelte'
 export { default as ColorPicker } from './ColorPicker.svelte'
 export { default as Error } from './Error.svelte'
@@ -10,7 +11,6 @@ export { default as Link } from './Link.svelte'
 export { default as LoadingScreen } from './LoadingScreen.svelte'
 export { default as Logo } from './Logo.svelte'
 export { default as Profile } from './Profile.svelte'
-export { default as ProfilePicture } from './ProfilePicture.svelte'
 export { default as ProgressBar } from './ProgressBar.svelte'
 export { default as QR } from './QR.svelte'
 export { default as RecoveryPhrase } from './RecoveryPhrase.svelte'
